Remove legacy Lambda ESM v1 feature (#11733)
