--- localstack-core/localstack/services/lambda_/event_source_listeners/exceptions.py
@@ -1,5 +1,0 @@
-"""Exceptions for lambda event source mapping machinery."""
-
-
-class FunctionNotFoundError(Exception):
-    """Indicates that a function that is part of an existing event source listener does not exist."""
