+++ tests/aws/services/lambda_/event_source_mapping/utils.py
@@ -1,6 +1,3 @@
-from localstack.config import LAMBDA_EVENT_SOURCE_MAPPING
-from localstack.testing.aws.util import is_aws_cloud
-
 _LAMBDA_WITH_RESPONSE = """
 import json
 
@@ -13,11 +10,3 @@ def handler(event, context):
 def create_lambda_with_response(response: str) -> str:
     """Creates a lambda with pre-defined response"""
     return _LAMBDA_WITH_RESPONSE.format(response=response)
-
-
-def is_v2_esm():
-    return LAMBDA_EVENT_SOURCE_MAPPING == "v2" and not is_aws_cloud()
-
-
-def is_old_esm():
-    return LAMBDA_EVENT_SOURCE_MAPPING == "v1" and not is_aws_cloud()
